requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
httpx==0.27.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
